requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0

# Database
psycopg2-binary==2.9.9
python-decouple==3.8

# Celery
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0

# Scraping & LLM
stagehand
playwright
openai==1.3.0
anthropic==0.8.0

# Data Processing
pandas==2.1.3
python-dateutil==2.8.2

# Development
ipython==8.17.2
django-extensions==3.2.3

# Production
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.38.0